Created a test for snowflake_connection_details method. Add GitHub actions workflow CI test.
