+++ .env_test
@@ -1,0 +1,6 @@
+SNOWFLAKE_USER=A
+SNOWFLAKE_ROLE=B
+SNOWFLAKE_PASSWORD=C
+SNOWFLAKE_ACCOUNT=D
+SNOWFLAKE_REGION=E
+SNOWFLAKE_WAREHOUSE=F
