+++ .github/workflows/on_push.yml
@@ -1,0 +1,55 @@
+name: run python tests
+
+on: push
+
+jobs:
+  run-unit-tests:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: checkout-code
+      uses: actions/checkout@v2
+
+    - name: cache-python-requirements
+      uses: actions/cache@v2
+      id: pip-cache
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
+        restore-keys: |
+          ${{ runner.os }}-pip
+
+    - name: setup-python
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.8
+
+    - name: install-python-requirements
+      run: |
+        python -m pip install --upgrade pip
+        pip install --requirement requirements.txt
+
+    - name: run-unit-tests
+      run: |
+        pytest -vv .
+  #
+  # run-snowflake-builder:
+  #   runs-on: ubuntu-latest
+  #   env:
+  #     SNOWFLAKE_USER: ADAM
+  #
+  #     steps:
+  #   - name: create-dev-environment
+  #     env:
+  #       environment: dev
+  #     run: python ./code/python/snowflake/build_account.py
+  #
+  #   - name: create-staging-environment
+  #     env:
+  #       environment: staging
+  #     run: python ./code/python/snowflake/build_account.py
+  #
+  #   - name: create-prod-environment
+  #     env:
+  #       environment: prod
+  #     run: python ./code/python/snowflake/build_account.py
