+++ snowflake_connector.py
@@ -1,6 +1,7 @@
 import snowflake.connector
 from snowflake.connector import DictCursor
 import pandas as pd
+import os
 
 
 class SnowflakeConnector:
@@ -54,3 +55,16 @@ class SnowflakeConnector:
 
     def closer_cursor(self, cursor):
         cursor.close()
+
+
+def snowflake_connection_details():
+    snowflake_connection_details = {
+        "user": os.environ.get("SNOWFLAKE_USER"),
+        "role": os.environ.get("SNOWFLAKE_ROLE", "SYSADMIN"),
+        "password": os.environ.get("SNOWFLAKE_PASSWORD"),
+        "account": str(os.environ.get("SNOWFLAKE_ACCOUNT"))
+        + "."
+        + os.environ.get("SNOWFLAKE_REGION", "eu-west-1"),
+        "warehouse": os.environ.get("SNOWFLAKE_WAREHOUSE", "COMPUTE_WH"),
+    }
+    return snowflake_connection_details
